+++ src/Payout.cpp
@@ -14,14 +14,14 @@ Payout::Payout() {
 
 Payout::~Payout() {}; // Don't know if we need this for payout
 
-int Payout::calculatePayout(int playerScore, int betAmout) {
+double Payout::calculatePayout(int playerScore, double betAmout) {
 
     auto it = mapPayout.find(playerScore);
 
     if (it != mapPayout.end()) {
         
         double multiplayer = it -> second;
-        int payout = betAmout * multiplayer;
+        double payout = betAmout * multiplayer;
 
         std::cout << "The payout Amount" << payout << std::endl;
